LayoutTests/imported/w3c/web-platform-tests/service-workers/service-worker/fetch-mixed-content-to-outscope.https-expected.txt

CONSOLE MESSAGE: line 25: The page at https://localhost:9443/service-workers/service-worker/fetch-mixed-content-to-outscope.https.html was allowed to display insecure content from http://localhost:8800/service-workers/service-worker/resources/fetch-access-control.py?PNGIMAGE.

CONSOLE MESSAGE: FetchEvent.respondWith received an error: TypeError: Not allowed to request resource
CONSOLE MESSAGE: Cannot load http://localhost:8800/service-workers/service-worker/resources/fetch-access-control.py?PNGIMAGE.
CONSOLE MESSAGE: line 42: The page at https://localhost:9443/service-workers/service-worker/fetch-mixed-content-to-outscope.https.html was allowed to display insecure content from http://127.0.0.1:8800/service-workers/service-worker/resources/fetch-access-control.py?PNGIMAGE.

CONSOLE MESSAGE: FetchEvent.respondWith received an error: TypeError: Not allowed to request resource
CONSOLE MESSAGE: Cannot load http://127.0.0.1:8800/service-workers/service-worker/resources/fetch-access-control.py?PNGIMAGE.


PASS Verify Mixed content of fetch() in a Service Worker 
